[CHANGE] Refactor shard assignment from service to Project
There are cases when a single service contains enough projects that it
outgrows a single shard. Since service.name is unique, we could not use
the same service.name on multiple shards.

Refactoring Shard assignment to a Project gives us more flexibility for
managing our Projects, and changes Service to a grouping of related rules
and notifications

- Service.shard is moved to Project.shard
- Service block groups Projects by Shard
- Service list gets pagination. It's expected that in the default case,
  users will subscribe to their projects updates, and therefore see them
  on the home page