Split up bootstrap command for reliability
The bootstrap command has an issue where it is difficult to both
configure the settings and test at the same time, because there is no
easy way to also reload the django settings. This means we need to
separate the commands. We also want to run some tests that require the
database, but we can't have them enabled by default or other commands
like migrate will be stuck in a chicken/egg problem. To get around this
we create a decorator to check for database access for our checks that
require it.