Rule form block: fix spacing after "Enable" toggle
The ON/OFF toggle button for enabling the rule is rendered as an inline-block
element. Due to that, the next `<br>` element is acting as a line break, without
leaving an empty line separating the two fields.

Two options were considered in order to fix this issue:
 - Adding another `<br>` element to effectively create an empty line between the
   two fields.
 - Displaying the ON/OFF toggle button as a block element instead of as an
   inline-block element.

The second option was preferred because using `<br>` to visually separate
elements is not a good practice. CSS properties like `margin` or `padding` are
meant for that purpose.