Rule form block: fix spacing before "Description"
The normal space between fields on this form is 20 pixels, created by an empty
line using the `<br>` element.

However, the spacing before the "Description" fields was actually 30 pixels
because the previous element (a `<p>` holding a help text) had a `margin-bottom`
of 10 pixels, applied by the `help-block` class from Bootstrap.