Rule form block: fix rendering for "Clause" header
The header for the "Clause" section was rendered different than other headers.
It was rendered in bold text and with a colon at the end. This was caused by the
`label_tag` renderer, which is meant to be used for form fields such as `input`,
hence the bold text and the colon.

Using the `name` renderer along with a `title` filter for capitalizing the first
character rendered the "Clause" header in the same way as the other ones.