Rule form block: fix spacing between elements in "Clause" section
The use of multiple `<div class="body-panel">` within this section was creating
spaces between the elements that were way too big. Each one of those body panels
has a `padding-top` and a `padding-bottom` of 15 pixels applied by Bootstrap.
That means having two consecutive body panels creates an spacing of 30 pixels
between them.

In our case, we had three consecutive body panels:
 - body panel to contain the elements before the text area
 - body panel to contain the text  area
 - body panel to contain the "Test Query" button

The first body panel had a `<dl>` element at the end of it. Taking into account
that Bootstrap applies a 20 pixels `margin-bottom` to `<dl>` elements, and body
panel themselves have a separation of 30 pixels between them (as we already said
at the beginning), that sums up a total separation of 50 pixels.

The separation between the text area and the "Test Query" button was 30 pixels,
created by the two consecutive body panels.

In order to fix this and have a consistent separation between all the elements,
a single body panel was using to contain all the elements. For the special case
of the separation between the text area and the "Test Query" button, a `<br>`
element was added (despite of being a bad practice) to keep it consistent with
the rest of the code.