Refresh the Home page when it's empty
Before, when the Home page was empty it just included a small text linking to
the services and data sources.

Now it has a more friendly headline, a bigger font to improve readability, and a
new link for directly creating a service. This new link will save the user from
having to navigate to the "Services" page in order to create its first service.