templates/promgen/rule_detail: Show description
Before, when you wanted to see the description of a rule, you needed to edit it
in order to load the form and then look at the description field. That was a
little bit weird and could be considered as an anti pattern for good usability.

Now, we have added the description to the rule detail page. Since the
description is an optional field, it is displayed conditionally.