Update test_alert_rules
Rewrite our alert rule tests to use our fixtures. This means that we no
longer need to do as much setup in the individual test cases and can
focus more on the specific logic we're testing.

Because we're using a rule from a fixture, we also need to update our
test template a bit.