Rule Test should check for unrelated labels
A common source of confusion when writing Rules, is the global namespace
of Prometheus when on a shared server. It's very easy to write a rule
that covers more projects than expected.

To help avoid this, we want to upgrade the rule test button to check the
labels that are returned vs the expected labels.