Stricter checks about silencing global rules
Until we have a more robust permission system, we need to be cautious
and ensure that users can not accidentally set a global level silence.
To do this, we will check to see if the pending silence has alertname as
a matcher, and if so, enforce that it also includes a service or project
label.