Create SilenceErrors enum to better organize expected errors
Django's ValidationError supports a code= parameter that we can use in
our test cases to avoid checking on string alone. We can use a custom
ValidationEnum to add an error property that returns our enum formatted
as a Django ValidationError. This allows us to refer to errors in
various places using a namespaced error code.