[BUGFIX] Fix for rule prefetch on /search
In f48218e3228ab625362c66811c11597381f0a872 we refactored how rules
stored labels/annotations but when we updated this prefetch entry we
accidentally changed it so that is no longer a proper tuple. This needs
to be a tuple so that it does not try to treat each character as an
entry.

(cherry picked from commit 2f1674b5d99d0f5221239078d5e083dd27f75fd6)