vue: Convert 'urlize' filter into a method
Filters are removed from Vue 3 and no longer supported, so in order to make the
future Vue 3 migration simpler we are going to convert them into methods, as
recommended [1].

Given those filters could be used by multiple components, we are going to store
them as mixins instead of defining them in specific components, which could lead
to code duplication.

Note that the only places where we were using this 'urlize' Vue filter was
inside an ':inner-html.prop' binding which seems undocumented nowadays and only
works with filters, not with methods [2]. We have replaced it with the `v-html`
built-in directive [3] which does work with methods and will continue to work in
Vue 3.

1: https://v3-migration.vuejs.org/breaking-changes/filters.html
2: https://stackoverflow.com/a/49868526
3: https://vuejs.org/api/built-in-directives.html#v-html