Migrate to Vue 3
In order to have better control of the version we are using, we have
placed the Vue import into a dedicated folder that includes the version
on its name.

In Vue 3, the devtools are not enabled through 'Vue.config'. They are
always enabled in the development build of Vue 3, and disabled in the
production one.
From now on, when running Promgen in debug mode we will use the
development build of Vue 3. When running Promgen in production mode we
will use the production build of Vue 3, as recommended [1].

In Vue 3, variables that are meant to be reactive (Vue tracks them and
updates the DOM whenever their value changes) need to be created making
use of the 'ref' or 'reactive' functions.

In Vue 3, apps are created with 'Vue.createApp' instead of 'new Vue'.

In Vue 3, 'Vue.set' does not exist, so all the code making use of that
has been rewritten appropriately.

In Vue 3, whitespace characters between elements that contain newlines
are removed by default. Because of that, we need to use the 'whitespace'
compiler option to change that behaviour, otherwise our UI elements such
as buttons will be rendered next to each other without any space in
between.

In Vue 3, components are not created using 'Vue.component'. Instead, the
application instance provides a 'component' method for registering
app-scoped components.

We have moved the instruction for mounting the Vue app at the end of the
'base.html' file to ensure all the necessary DOM elements are ready.

The 'exporter-test-result' component was being mounted on demand using a
'target' property, and also made visible by changing CSS properties from
JavaScript.
That has been replaced by directly placing the component element in the
right location and controlling its visibility with a 'show' property.

1: https://vuejs.org/guide/best-practices/production-deployment.html