settings: Move INTERNAL_IPS config to top level
When running in debug mode, it is necessary to have the INTERNAL_IPS variable
properly set in order to have some functionality.

Before we were assuming that it was only necessary for the Django toolbar,
however, according the documentation [1] it is also necessary for exposing some
variables in the template context, which is something we are going to need in
the future.

1: https://docs.djangoproject.com/en/3.2/ref/settings/#internal-ips