Vue: Add support for Pinia
This will allow us to define stores using Pinia, which is the new
standard for Vue stores replacing Vuex.
So far we just import the library and make the Vue app use Pinia.

The idea is to define the stores in separate JS files living inside a
stores directory, which will be more organised than our current state
where all the stores are defined in a single JS file among other things.