Vue: Migrate silence store to Pinia
This will make the future transition to a single file component much
easier. Also, the state is now defined on its own file so the
promgen.vue.js file becomes smaller, and at the same time the source
code structure is more organized.

Note we have modified the Makefile because the build of the static files
now not only depends on changes made to main.js, but anything inside the
src directory.