Vue: Migrate silence form to single-file component
Single file components are easier to manage, and on top of that we can
easily import libraries, other components and stores, define scoped CSS
styles, etc.

Now all the code for that component is in a single place, rather than
having the JS code in one place and the HTML template in another one.