Vue: Add support for vue-i18n
This will allow us to do translations on the front end without relying
on Jinja templating. This is important because Jinja templating will not
work in single file components.

We add support for English and Japanese locales, which is what Promgen's
Django side currently supports. The locale in Django is chosen based on
the browser's configuration for the preferred language for displaying
pages. We have done the same in the Vue side.