Vue: Migrate exporter test result store to Pinia
This will make the future transition to a single file component much
easier. Also, the state is now defined on its own file so the
promgen.vue.js file becomes smaller, and at the same time the source
code structure is more organized.

Since this store is going to be used by multiple components related to
exporter tests, we have renamed it to something more appropriate:
exporter test.