Vue: Migrate exporter result to single-file component
Following what we did for the silence form, we are now migrating the
exporter result to a single-file component. This time we have also
changed the name to something more appropriate: exporter test result.

The previous component had a "results" prop that was unused, since it was
already provided by the store. We have removed it in the new component.

Also, we have moved the visible state to the store, so the component is
a little bit simpler.