Vue: Migrate promql query to single-file component
Following what we did for the silence form, we are now migrating the
promql query to a single-file component.

The "localize" and "percent" global filters that were previously defined
as mixins have been moved into a "filters.js" file under "src/utils", so
they can be imported individually as needed.

The component visibility that was previously controlled by directly
changing the element's "display" property to "inline" is now being
controlled by a "visible" reactive variable used in a "v-show"
directive.