Add notifiers to audit logs
We had a query for listing notifiers in views.py::AuditList, however, we never
registered any logs from signals.py, therefore the result of that query was
always empty.

This commit adds the necessary instructions for registering logs when creating,
updating or deleting notifiers, so they will appear in the audit logs.

Since the data is dumped using model_to_dict(), we also needed to modify
models.py to ensure that sensible information is not displayed. That is, when
the users have set the "alias" field because they want to hide the value.