util: New post_with_retry function
This function is a thin wrapper of the existing util::post function but with a
retry mechanism that depends on the response's status code.

This function can be used by our notifiers so we can retry sending a
notification if, for instance, we have received a "Too Many Requests (429)" the
first time.