Add UI feedback to notifier test button
After clicking the notifier test button, it can take some time until the request
is processed, the response is received and the whole form submission ends up on
a page refresh.

Because of that potential delay, the users may think that they did not click the
button properly and try to click it again.

To avoid that situation we have added some UI feedback so when the users click
the test button, it gets disabled so they cannot click it again, and also its
text changes to "Testing..." to clearly state that the test is ongoing.

The corresponding localization for "Testing" has also been added.