templates: Add v-pre to protect against XSS
When rendering the front end code with Jinja on the server side, the resulting
rendered code may include Vue delimiters that will be interpreted by Vue on the
client side.

To prevent that, we have added v-pre to some elements in those places where we
think we can have issues.