views/NotifierUpdate: Escape square brackets in global message
Despite of having added v-pre in the right place, for some reason it looks like
that template is not being updated properly when re-running the Django app.

Until we find a fix for that issue we are adding a simple ugly escape for square
brackets to tackle the XSS problem.