Move promql-query request to the backend
The API end point of some shards are protected with HTTP Basic Auth. This is
making the promql-query component to fail receiving a '401 Unauthorized' error
response.

Now that the Shard model supports an "authorization" field to specify the value
of the "Authorization" HTTP header, we can address this issue.

However, if we perform the request in the front end, that information would be
visible by a simple inspection of the HTTP headers. Because of that we have
moved it to the back end.