PrometheusProxy: Add support for authorization
Now that our Shard model supports an "authorization" field to specify the value
of the HTTP "Authorization" header, we can use it in our PrometheusProxy class.

Note that we have replace "host" with "shard" since the objects obtained by the
'models.Shard.objects.filter(proxy=True)' instruction are shards, not hosts.