Test metrics validity when testing Exporter
When using the 'test' button for an exporter, we previously *only*
checked the status code and assumed things were good. It is very possible
for an incorrect endpoint to be registered, so we should ensure that the
registered endpoint produces correct Prometheus metrics.

Right now, we will return a somewhat *developer* like output, but I
expect to make the output more user friendly as we cleanup some of our
internal test URLs.