Add renderer test cases
In the future, I would like to improve the way we handle some of the
rendering, so the first step is to ensure we have some proper test cases
to check for when rendering changes.

Most of this is updates for test cases, but also add a minor new route
as part of a future cleanup of the public api routes.