Convert silence_row into a Vue component
This will help with future development around the Silence area.

The component allows choosing the color of the matchers, and also emits an event
when a matcher is clicked. That way we can reuse the component for multiple
purposes, displaying the matchers in a different color and doing an different
action when a matcher is clicked.

A few bits had to be added in order to maintain the current style, such as the
template and tbody tags.