Service button group (Part 1)
In the interest of consolidating some of the UI, we want to move more
actions under an actions dropdown.

For Part 1, we move a number of menu options under a new
service_action_button_group, and then unify our service_detail and
service_list pages. In a future pull request, we will do more changes to
how we display things under a service_block