Add Permission Checks for Promgen Web
We have created a mixin class to check the permissions of a logged-in user for a View class.
A user will be allowed to access the View if they have at least one of the permissions declared
in the View class for the specific object corresponding to that View class. Since we only define
permissions for Service, Project, and Group, other objects will have their permissions checked
on the parent object.