Add 'Permissions' panel to UI
Add a panel to the UI on each detail screen of Service/Project/Farm object
to assign or remove roles for users. User needs to have the Admin role
to assign or remove roles for users through this panel.
However, the permission of the owner always is ADMIN. User cannot change
or remove the owner's permission without transferring the ownership.