service_detail: Display active silences
This feature was removed by commit 1c7574d when we modified the service detail
page to use tabs.

At that time, we were already displaying the list of silences in a modal dialog,
so I guess we expected that the users would go there and search for their
silences filtering by service, therefore displaying the active silences in the
service display page was no longer necessary.

However, that means they will need to explicitly search for their silences if
they want to know if there are any active silences, otherwise they will not
notice.

Displaying the active silences for a service right away is a useful feature we
want to have back, so this commit adds a new button that, when clicked, it will
reuse the silences modal dialog to display only the ones for that service.
If there are no active silences for that service, the button will not be there.