service_detail: Display active alerts
This feature was removed by commit 1c7574d when we modified the service detail
page to use tabs, probably by mistake. At that point the list of active silences
for the service was also removed, so we probably deleted too many lines without
noticing.

We still want to display the active alerts for a given service on its service
detail page, because trying to find them in the global list of alerts can be a
nightmare.