Change the operator to avoid misleading when creating new silence
In the filter feature of the silence list form, the =~ operator caused confusion as users thought
they could use regex. However, only values ending in * are regex, others are considered literals.
Therefore, for the sake of avoiding confusion among our users, we change the UI to use
the = operator. It is better to put an = operator in there until we add support for new operators
in the future. We know this is not 100% accurate, because anything ending in * is considered
a regular expression, therefore the actual operation is =~ in that case.
We also change a misleading comment in the code that led to the addition of the =~ operator
in the UI.