Change the operator to avoid misleading when filtering on the global list of silence
Currently, when using the filter feature on the global list of silence, we are using
the === operator in JavaScript. The values being compared are also entirely chosen from
a list of specific values. Therefore, the operator that should be displayed here is = instead of =~.
We have made this change to avoid misunderstandings.