Trim whitespaces of search input
Sometimes, when users want to search for a keyword, it may accidentally include leading
or trailing spaces because the user copied the keyword from elsewhere or due to a typing
error. This results in empty search results because nothing matches the input.
Therefore, before starting the search, we clean up the user's input by trimming the
spaces at the beginning and at the end of the input string.