Use DRF's APIView for AlertReceiver API
Django's View class is designed for building traditional web applications that render
HTML templates. On the other hand, APIView is part of Django REST Framework and is specifically
designed for building robust RESTful APIs. Therefore, we have transitioned from using
Django's View class to Django REST Framework's APIView for the AlertReceiver API. This
change allows us to implement advanced features such as authentication and permissions
in the future.
This update does not affect the current behavior of the AlertReceiver API.