New template for Farm registering
We have created a separate template named farm_register.html specifically for
the case of creating a new Farm. Currently, to avoid creating a new Farm not linked to
any Project, a Farm can only be created through the Farm Register button on
the Project Detail Page. To facilitate the creation of a new Farm, we allow users
to enter hosts at the same time when creating the Farm. The old template will be
renamed to farm_update.html to match its intended use.