Replace the Vue component used for displaying data source's load
We have replaced the API used by the "promql-query" Vue component, which was responsible
for querying data about samples and exporters of the Shard, with a new API. The original
API, promql-query, allowed users to input PromQL as a parameter and returned the query
results in the API response. This capability led to a potential security issue where users
could copy, modify, and execute requests with queries different from those intended on
Promgen Web. To prevent such exploits, we introduced a new API that accepts only a metric
parameter, enabling the client to choose between querying "samples" and "exporters." The
corresponding queries are now mapped on the backend to prevent manipulation.
Since the "promql-query" Vue component is no longer used elsewhere, we have renamed it to
"data-source-usage" to better reflect its purpose.

Following this change, the PromQL Query API became redundant. Consequently, we have
removed this API along with its associated View class.