Always display the data source's load
When registering a new Project we are presented with a list of data sources, and we
need to select one of them. To keep the data source's load balanced, we display a load
percentage and expect the users to select one of the data sources with a low load.

Currently, the code for displaying that load percentage only runs if the proxy feature
for that data source is enabled, even if the proxy end point is not used at all for
that matter.

Therefore, we make some necessary changes to display the data source's load percentage
even if the proxy feature is disabled for the data source.