Preventing the browser from caching promgen.vue.js file
We added a version number as a query string to the promgen.vue.js file URL to prevent
the user's browser from caching it. In the future, this version number will be
manually increased when we modify the promgen.vue.js file. This ensures that users
always receive the latest version of the JavaScript file every time we change it.