Improve UI for Farm Register page
We made a few changes to the FarmRegister view class and its template to improve
the UI of the Farm Register page. The farm_register.html included two tables from
FarmForm and HostForm, which caused many display issues. Therefore, we created a
new form named FarmRegisterForm that contains all the fields from both forms to
resolve all these issues and ensure the template is consistent with other Promgen's
forms.