Fix UI quirks in the Farm Register page
There was a disorganized arrangement of fields in the Farm Register form. There is a gap
in the row containing the "owner" label. Additionally, the row containing the "hosts"
label and textarea is misaligned. Therefore, we placed all the fields in the same table
so they can be automatically aligned properly.