Fix external Farm linkage
After the changes that made the owner field in the Farm model not nullable [1],
linking an external Farm for the first time failed.

This is because when linking a Farm for the first time, a new Farm object is
created into the data base. However, the code that ran for that particular case
did not populate the owner at creation time, therefore it failed.

Now, when linking a Farm, first we try to get it from the DB. If it does not
exist, then we creat it with the owner field populated.

1: Commit d55dfc0

Tested-by: hoang <pham.ngoc.hoang@linecorp.com>