Remove the redundant contact point
After applying the Object-level-permissions authorization logic, the
Service/Farm/Project can only be viewed by its members, including the
owner. Additionally, we have the Members panel to view the list of members
and their emails, along with an icon to indicate who is the Owner. Therefore,
the "contact" label, which is the Owner, will become unnecessary. We have
removed this label.