Add migration file for assigning admin permission to the owner
After applying the Object-level-permissions authorization logic, the Service or Farm
or Project can only be viewed by its members, including the owner. This may confuse
users of Promgen instances that already have existing data because they will not be
able to see anything after the new feature is added. In such cases, the Promgen
administrator will need to manually reassign permissions to those users.

To partially support this task, we have added a migration script that automatically
assigns admin rights to the owners of each existing Service/Project/Farm.