Fix empty row when registering a new project
We are using the same form for registering and updating a project. Therefore, there is
an empty row at the end of the form when registering a new project because that row is
only for updating the service of the project. As a result, we have hidden this row
when registering a new project.