Not provide query results when testing Rule
From the Rule's edit page, users can write any query and execute with the test button.
After that, the results are presented on the screen. The query is executed on all data
sources, and for that Promgen provides the basic auth credentials configured in for
each data source. This capability led to a potential security issue where users can run
any query anonymously.

Therefore, we want to stop providing the query results to the user, and instead, just
say if the rule would fire or not.