Fix error messages showed when testing a site rule
When the administrator edits a site rule and tests the query, error messages about
unrelated_service and unrelated_project are always shown. These messages shouldn't
appear in this case because the site rule doesn't link to any service or project.
We have fixed this issue to hide these error messages in this case.