Make notifier's filters not editable if 'show_edit' is false
Currently, when displaying the notifier list, the Actions column will not be shown if
show_edit is false. However, the filter column is still displayed along with the Add
and Remove buttons when show_edit is false. For instance, when we display the Service
notifiers from the Project detail page, the Action column is hidden by using the show_edit
variable. However, even if the Actions column is hidden because the users are not supposed
to edit those notifiers, they are still able to add filters to it, which is an edit action
on the notifier.
We have changed it to only display the filters as labels and not show any actionable
buttons so that the users are not supposed to add, edit or remove filters of a notifier
if the value of 'show_edit' is false.