Show operators of Silence's matchers on UI
We have updated the silence_row.html file so that the list of silences now correctly
displays the matchers and their operators, instead of only showing the '=' operator as
before.
The '=' operator in the silence_list_modal.html file has also been replaced with ':'
to indicate that filtering is taking place with the name and value, without regard to
the matcher's operator.