Add 'Site Configuration' model
We have created a new model named 'Site Configuration' that allows the Administrator
to create and edit key-value pairs directly on the Admin Site. This is beneficial for
developing features that require the use of setting values that can be flexibly
changed in the future, such as throttling, feature flags, etc.