Redirect to the current active Tab after editing data
Currently, if users edit data on any Tab, the page will reload and show the first Tab of
that page. This affects the user experience because the users will have to repeatedly
select the initial Tab if they want to edit multiple data in that Tab. We have fixed this
problem by putting the hash of the Tab to the redirect URLs in these cases.