Improve the Edit History feature
With the addition of two new columns in commit 714e1fcd, querying audit logs for a
specific Service or Project and its child objects has become simpler and better.
We have updated the query_set in the AuditList class so that queries are performed only
on the Audit model. The result is still returning audit logs for both the main object
and its child objects, even if the child objects have been deleted.