Add Audit Logs when assigning or removing permission
We linked the UserObjectPermission model to the create_log and delete_log functions using
the post_save and post_delete signals to generate audit logs when permissions are assigned
to or removed from a user.
However, the post_save signal will be temporarily disconnected when migration file number
0029 is executed. In this scenario, Promgen assigns permissions automatically, so Audit
Logs are not needed.