Support change member's role on UI
Currently, the only way to change a member's role is to delete the member and add them
again with the desired role. From a UX perspective, this is somewhat inconvenient.
Therefore, we have added a small button next to each member's role label in the member list
table to quickly support changing the member's role.