Add Group pages to UI
We have made changes to the Promgen Web to help users interact with the features of
the Group model. Users can now access the list of Groups through a new menu at the top
bar. They can also access the Group detail page and perform actions such as managing
the member list or viewing the edit history of the Group there.